+++ myia/pipeline.py
@@ -1,279 +1,7 @@
 """Tools to generate and configure Myia's operation pipeline."""
 
 
-import inspect
-from .utils import TypeMap, Named
-
-
-# Use in a merge to indicate that a key should be deleted
-DELETE = Named('DELETE')
-
-
-# Placeholder for MergeMode.__init__
-_ABSENT = Named('_ABSENT')
-
-
-class MergeMode:
-    """Wraps a value to control how it is merged.
-
-    Class attributes:
-        mode: The merge mode to use.
-
-    Attributes:
-        value: The value to merge.
-
-    """
-
-    mode = 'merge'  # NOTE: This is the default merge mode used by merge()
-
-    def __init__(self, __value=_ABSENT, **kwargs):
-        """Initialize a MergeMode.
-
-        The value to merge will be the single argument given, or if there is no
-        positional argument, the keywords dictionary.
-        """
-        if __value is _ABSENT:
-            self.value = kwargs
-        else:
-            assert not kwargs
-            self.value = __value
-
-
-class Merge(MergeMode):
-    """Merge normally."""
-
-    mode = 'merge'
-
-
-class Override(MergeMode):
-    """Do not concatenate sequences."""
-
-    mode = 'override'
-
-
-class Reset(MergeMode):
-    """Throw away the previous value."""
-
-    mode = 'reset'
-
-
-_cleanup_map = TypeMap()
-
-
-@_cleanup_map.register(object)
-def _cleanup_object(value):
-    return value
-
-
-@_cleanup_map.register(MergeMode)
-def _cleanup_MergeMode(mm):
-    return mm.value
-
-
-@_cleanup_map.register(dict)
-def _cleanup_dict(d):
-    return type(d)({k: cleanup(v) for k, v in d.items() if v is not DELETE})
-
-
-def _cleanup_sequence(xs):
-    return type(xs)(cleanup(x) for x in xs)
-
-
-_cleanup_map.register(tuple, _cleanup_sequence)
-_cleanup_map.register(list, _cleanup_sequence)
-_cleanup_map.register(set, _cleanup_sequence)
-
-
-def cleanup(x):
-    """Remove all MergeMode and DELETE instances from the data."""
-    return _cleanup_map[type(x)](x)
-
-
-_merge_map = TypeMap(discover=lambda cls: getattr(cls, '__merge__', None))
-
-
-@_merge_map.register(dict)
-def _merge_dict(d1, d2, mode):
-    if mode == 'reset':
-        return type(d1)(d2)
-
-    rval = type(d1)()
-    for k, v in d1.items():
-        if k in d2:
-            v2 = d2[k]
-            if v2 is DELETE:
-                pass
-            else:
-                rval[k] = merge(v, v2, mode)
-        else:
-            rval[k] = v
-    for k, v in d2.items():
-        if k not in d1:
-            rval[k] = cleanup(v)
-    return rval
-
-
-@_merge_map.register(tuple)
-def _merge_tuple(xs, ys, mode):
-    xs = cleanup(xs)
-    ys = cleanup(ys)
-    if mode == 'merge':
-        return xs + ys
-    else:
-        return ys
-
-
-@_merge_map.register(list)
-def _merge_list(xs, ys, mode):
-    xs = cleanup(xs)
-    ys = cleanup(ys)
-    if mode == 'merge':
-        return xs + ys
-    else:
-        return ys
-
-
-@_merge_map.register(set)
-def _merge_set(xs, ys, mode):
-    xs = cleanup(xs)
-    ys = cleanup(ys)
-    if mode == 'merge':
-        return xs | ys
-    else:
-        return ys
-
-
-@_merge_map.register(object)
-def _merge_object(a, b, mode):
-    return cleanup(b)
-
-
-def merge(a, b, mode=MergeMode.mode):
-    """Merge two data structures.
-
-    Arguments:
-        a: The original data.
-        b: The data to merge.
-        mode:
-            'merge': (default) Sequences will be concatenated, sets merged,
-                and dictionaries merged according to their keys.
-            'override': Dictionaries are merged, but sequences are not
-                concatenated.
-            'reset': b is returned, or takes primacy over the data in a.
-    """
-    if isinstance(b, MergeMode):
-        mode = b.mode
-        b = b.value
-    assert not isinstance(a, MergeMode)
-    return _merge_map[type(a)](a, b, mode)
-
-
-def _filter_keywords(f, kw):
-    spec = inspect.getfullargspec(f)
-    if spec.varkw:
-        return kw, {}
-    valid = spec.args + spec.kwonlyargs
-
-    good = {k: v for k, v in kw.items() if k in valid}
-    bad = {k: v for k, v in kw.items() if k not in valid}
-
-    return good, bad
-
-
-class NS:
-    """Namespace for pipeline stages.
-
-    This namespace preserves key order in the representation, unlike
-    types.SimpleNamespace.
-    """
-
-    def __init__(self, **kwargs):
-        """Initialize NS."""
-        self.__dict__.update(kwargs)
-
-    def __getitem__(self, item):
-        return self.__dict__[item]
-
-    def __setitem__(self, item, value):
-        self.__dict__[item] = value
-
-    def __repr__(self):
-        args = [f'{k}={v}' for k, v in self.__dict__.items()]
-        return f'NS({", ".join(args)})'
-
-
-class Partial:
-    """Partial application of a function.
-
-    This differs from functools.partial in a few ways:
-
-    * Only keyword arguments are accepted.
-    * Argument names are validated immediately.
-    * It is possible to merge two partials, with the second updating the
-      parameters of the first. It is also possible to merge a dict and
-      a Partial. Merge and Override work basically like on dictionaries.
-    * `merge(partial1, Override(partial2))` lets us change the
-      constructor.
-    """
-
-    def __init__(self, func, **keywords):
-        """Initialize a Partial."""
-        self.func = func
-        self.keywords = keywords
-        self._validate()
-
-    def _validate(self):
-        """Check that all the argument names are valid."""
-        if isinstance(self.func, type):
-            f = getattr(self.func, '__init__', self.func)
-        else:
-            f = self.func
-        _, invalid = _filter_keywords(f, self.keywords)
-        if invalid:
-            keys = ", ".join(f"'{k}'" for k in invalid.keys())
-            raise TypeError(f"{f} has no argument(s) named {keys}")
-
-    def partial(self, **keywords):
-        """Refine this Partial with additional keywords."""
-        return merge(self, keywords)
-
-    def __call__(self, **kwargs):
-        """Merge stored arguments with kwargs and call the function."""
-        return self.func(**self.keywords, **kwargs)
-
-    def __merge__(self, partial, mode):
-        """Combine arguments from two partials."""
-        if isinstance(partial, dict):
-            partial = Partial(self.func, **partial)
-
-        assert isinstance(partial, Partial)
-
-        if partial.func is self.func \
-                or mode == 'override' or mode == 'reset':
-            func = partial.func
-        else:
-            raise ValueError('Incompatible func')
-
-        kwargs = merge(self.keywords, partial.keywords, mode)
-
-        return Partial(func, **kwargs)
-
-    def __repr__(self):
-        args = [f'{k}={v}' for k, v in self.keywords.items()]
-        return f'{self.func.__name__}({", ".join(args)})'
-
-
-class Partializable:
-    """Class for which partial instances may be created.
-
-    This defines `Class.partial(arg=val, ...)`, which is equivalent to
-    `Partial(Class, arg=val, ...)`.
-    """
-
-    @classmethod
-    def partial(cls, **kwargs):
-        """Return a Partial on this class constructor."""
-        return Partial(cls, **kwargs)
+from .utils import merge, Merge, NS, Partial, Partializable, partition_keywords
 
 
 class PipelineDefinition:
@@ -496,7 +224,7 @@ class _PipelineSlice:
     def __call__(self, **args):
         for step in self.pipeline._seq[self.slice]:
             if step.active:
-                valid_args, rest = _filter_keywords(step.step, args)
+                valid_args, rest = partition_keywords(step.step, args)
                 args = {**args, **step.step(**valid_args)}
         return args
