+++ myia/utils/misc.py
@@ -301,6 +301,30 @@ class Events:
             setattr(self, event_name, ev)
 
 
+class NS:
+    """Simple namespace, acts a bit like a dict with dot access on keys.
+
+    This is different from Namespace below.
+
+    This namespace preserves key order in the representation, unlike
+    types.SimpleNamespace.
+    """
+
+    def __init__(self, **kwargs):
+        """Initialize NS."""
+        self.__dict__.update(kwargs)
+
+    def __getitem__(self, item):
+        return self.__dict__[item]
+
+    def __setitem__(self, item, value):
+        self.__dict__[item] = value
+
+    def __repr__(self):
+        args = [f'{k}={v}' for k, v in self.__dict__.items()]
+        return f'NS({", ".join(args)})'
+
+
 class Namespace:
     """Namespace in which to resolve variables."""
