+++ tests/test_pipeline.py
@@ -1,8 +1,7 @@
 
 import pytest
-from myia.pipeline import merge, NS, Partial, PipelineStep, \
-    PipelineDefinition, Merge, Reset, Override, DELETE, cleanup
-from myia.utils import TypeMap
+from myia.pipeline import PipelineStep, PipelineDefinition
+from myia.utils import Merge, Reset
 
 
 class OpStep(PipelineStep):
@@ -39,159 +38,6 @@ def op_pipeline():
             square=OpStep.partial(op=lambda p, x: x * x),
         )
     )
-
-
-def test_merge():
-
-    assert merge(1, 2) == 2
-    assert merge([1, 2], [3, 4]) == [1, 2, 3, 4]
-    assert merge((1, 2), (3, 4)) == (1, 2, 3, 4)
-    assert merge({1, 2}, {3, 4}) == {1, 2, 3, 4}
-
-    a = dict(a=1, b=2, c=dict(d=3, e=[4, 5]))
-    b = dict(a=1, b=2, c=dict(e=[6, 7], f=8))
-    c = dict(a=1, b=2, c=dict(d=3, e=[4, 5, 6, 7], f=8))
-
-    assert merge(a, b) == c
-
-    dlt = dict(c=DELETE, d=3)
-    assert merge(a, dlt) == dict(a=1, b=2, d=3)
-
-
-def test_merge_subclass():
-
-    tm = TypeMap({int: "int"})
-    mtm = merge(tm, {str: "str"})
-    assert isinstance(mtm, TypeMap)
-    assert mtm == TypeMap({int: "int", str: "str"})
-
-
-def test_merge_modes():
-
-    for x, y in [({1, 2}, {3, 4}),
-                 ([1, 2], [3, 4]),
-                 ((1, 2), (3, 4))]:
-
-        assert merge(x, y, mode='reset') == y
-        assert merge(x, Reset(y)) == y
-        assert merge(x, Reset(y), mode='merge') == y
-
-        assert merge(x, y, mode='override') == y
-        assert merge(x, Override(y)) == y
-        assert merge(x, Override(y), mode='merge') == y
-
-    a = {'a': 1}
-    b = {'b': 2}
-    c = {'a': 1, 'b': 2}
-
-    assert merge(a, b, mode='reset') == b
-    assert merge(a, b, mode='override') == c
-
-    a = {'a': [1, 2], 'b': [3, 4]}
-    b = {'a': [5, 6], 'b': Override([7, 8])}
-    c = {'a': [1, 2, 5, 6], 'b': [7, 8]}
-    d = {'a': [5, 6], 'b': [7, 8]}
-
-    assert merge(a, b) == c
-    assert merge(a, b, mode='override') == d
-
-
-def test_cleanup():
-    a = dict(a=1, b=[2, Merge(3)], c=Override(4), d=DELETE)
-    assert cleanup(a) == dict(a=1, b=[2, 3], c=4)
-
-
-def test_cleanup_subclass():
-    a = TypeMap({int: Merge("int")})
-    ca = cleanup(a)
-    assert isinstance(ca, TypeMap)
-    assert ca == TypeMap({int: "int"})
-
-
-def test_NS():
-    ns = NS(x=1, y=2)
-
-    assert ns.x == 1
-    assert ns.y == 2
-
-    ns.a = 3
-    assert ns.a == 3
-    assert ns['a'] == 3
-
-    ns['b'] = 4
-    assert ns['b'] == 4
-    assert ns.b == 4
-
-    assert repr(ns) == 'NS(x=1, y=2, a=3, b=4)'
-
-
-def test_Partial():
-
-    def f(x, y):
-        return x + y
-
-    p1 = Partial(f, x=10)
-    p2 = Partial(f, y=20)
-
-    assert p1(y=3) == 13
-    assert p2(x=3) == 23
-    assert merge(p1, p2)() == 30
-    assert merge(p1, {'y': 20})() == 30
-
-    with pytest.raises(TypeError):
-        Partial(f, z=10)
-
-    p3 = Reset(Partial(f, y=20))
-    with pytest.raises(TypeError):
-        merge(p1, p3)()
-
-    p4 = Partial(f, x=[1, 2])
-    p5 = Partial(f, x=[3, 4])
-
-    assert merge(p4, p5)(y=[10]) == [1, 2, 3, 4, 10]
-    assert merge(p5, p4)(y=[10]) == [3, 4, 1, 2, 10]
-
-    p6 = Override(Partial(f, x=[3, 4]))
-    assert merge(p4, p6)(y=[10]) == [3, 4, 10]
-
-    def g(x, y):
-        return x * y
-
-    p7 = Partial(g, y=20)
-    with pytest.raises(ValueError):
-        merge(p1, p7)
-
-    p8 = Override(Partial(g, y=20))
-    assert merge(p1, p8)() == 200
-
-    def h(**kw):
-        return kw
-
-    p9 = Partial(h, x=10)
-    assert p9(y=20, z=30) == dict(x=10, y=20, z=30)
-
-    str(p9), repr(p9)
-
-
-def test_Partial_class():
-
-    class C:
-        def __init__(self, x, y):
-            self.x = x
-            self.y = y
-
-        def __call__(self):
-            return self.x + self.y
-
-    p1 = Partial(C, x=10)
-    p2 = Partial(C, y=20)
-
-    assert p1(y=3)() == 13
-    assert p2(x=3)() == 23
-    assert merge(p1, p2)()() == 30
-
-    with pytest.raises(TypeError):
-        Partial(C, z=10)
 
 
 def test_PipelineDefinition_index(op_pipeline):
